Fixes #813: Add parseDate function as a helper function
