+++ src/components/SkillFeedbackCard/SkillFeedbackCard.js
@@ -24,6 +24,8 @@ import EditBtn from 'material-ui/svg-icons/editor/mode-edit';
 // CSS
 import './SkillFeedbackCard.css';
 
+import { parseDate } from '../../Utils/helperFunctions';
+
 const cookies = new Cookies();
 
 const iconButtonElement = (
@@ -89,42 +91,10 @@ class SkillFeedbackCard extends Component {
     this.handleDeleteClose();
   };
 
-  parseDate = dtstr => {
-    if (dtstr) {
-      // replace anything but numbers by spaces
-      dtstr = dtstr.replace(/\D/g, ' ');
-      // trim any hanging white space
-      dtstr = dtstr.replace(/\s+$/, '');
-      // split on space
-      var dtcomps = dtstr.split(' ');
-      // not all ISO 8601 dates can convert, as is
-      // unless month and date specified, invalid
-      if (dtcomps.length < 3) {
-        return 'Invalid date';
-      }
-      // if time not provided, set to zero
-      if (dtcomps.length < 4) {
-        dtcomps[3] = 0;
-        dtcomps[4] = 0;
-        dtcomps[5] = 0;
-      }
-      // modify month between 1 based ISO 8601 and zero based Date
-      dtcomps[1]--;
-      const convdt = new Date(
-        Date.UTC(
-          dtcomps[0],
-          dtcomps[1],
-          dtcomps[2],
-          dtcomps[3],
-          dtcomps[4],
-          dtcomps[5],
-        ),
-      );
-      let time = convdt.toString();
-      time = time.split(' ').slice(1, 4);
-      time[1] = `${time[1]},`;
-      return time.join(' ');
-    }
+  formatDate = timestamp => {
+    timestamp = timestamp.split(' ').slice(1, 4);
+    timestamp[1] = `${timestamp[1]},`;
+    return timestamp.join(' ');
   };
 
   render() {
@@ -180,7 +150,7 @@ class SkillFeedbackCard extends Component {
                 <div>
                   <div>{data.email}</div>
                   <div className="feedback-timestamp">
-                    {this.parseDate(data.timestamp)}
+                    {this.formatDate(parseDate(data.timestamp))}
                   </div>
                 </div>
               }
@@ -219,7 +189,7 @@ class SkillFeedbackCard extends Component {
               <div>
                 <div>{data.email}</div>
                 <div className="feedback-timestamp">
-                  {this.parseDate(data.timestamp)}
+                  {this.formatDate(parseDate(data.timestamp))}
                 </div>
               </div>
             }
