+++ src/components/SkillPage/SkillListing.js
@@ -35,6 +35,7 @@ import VersionBtn from 'material-ui/svg-icons/action/history';
 import ReactTooltip from 'react-tooltip';
 import colors from '../../Utils/colors';
 import urls from '../../Utils/urls';
+import { parseDate } from '../../Utils/helperFunctions';
 
 import './SkillListing.css';
 
@@ -520,41 +521,6 @@ class SkillListing extends Component {
     this.setState({ showAuthorSkills: false });
   };
 
-  parseDate = dtstr => {
-    if (dtstr) {
-      // replace anything but numbers by spaces
-      dtstr = dtstr.replace(/\D/g, ' ');
-      // trim any hanging white space
-      dtstr = dtstr.replace(/\s+$/, '');
-      // split on space
-      var dtcomps = dtstr.split(' ');
-      // not all ISO 8601 dates can convert, as is
-      // unless month and date specified, invalid
-      if (dtcomps.length < 3) {
-        return 'Invalid date';
-      }
-      // if time not provided, set to zero
-      if (dtcomps.length < 4) {
-        dtcomps[3] = 0;
-        dtcomps[4] = 0;
-        dtcomps[5] = 0;
-      }
-      // modify month between 1 based ISO 8601 and zero based Date
-      dtcomps[1]--;
-      const convdt = new Date(
-        Date.UTC(
-          dtcomps[0],
-          dtcomps[1],
-          dtcomps[2],
-          dtcomps[3],
-          dtcomps[4],
-          dtcomps[5],
-        ),
-      );
-      return convdt.toUTCString();
-    }
-  };
-
   testExample = (e, exampleText) => {
     let link = 'https://chat.susi.ai/?testExample=' + exampleText;
     window.open(link, '_blank');
@@ -765,7 +731,7 @@ class SkillListing extends Component {
                   <div>
                     <br />
                     Last modified at -
-                    {` ${this.parseDate(this.state.last_modified_time)}`}
+                    {` ${parseDate(this.state.last_modified_time)}`}
                   </div>
                 </div>
               </div>
